DE's standard crossover implemented: DE+APM and DE+TS.

+++ src/c/deleqc-ii/funcoes_teste.h
@@ -5,7 +5,6 @@
 void Problema_Teste(int prob, int *Restricoes, int *N_Coordenada, double *solucao, double ***bounds, int nCoordenadasNova, int nRestricoesNova, int* nGs, int* nHs);
 void Avalia(double *id, int N_Coordenada, int prob, double **A, double** bounds, double epsilonEqualityConstraints);
 void boundConstraints(double *id, int dimension, int nGs, int nHs, double** bounds);
-void sumViolation(double *individual, int dimension, int nGs, int nHs, double** bounds, double nLHs, double* maxConstraints);
+void sumViolation(double *individual, int dimension, int nGs, int nHs, double** bounds, int nLHs, double* maxConstraints);
 void linearConstraintsData(int prob, double** e, double** eTransposed, double* b);
-
